+++ packages/types/src/actions.ts
@@ -18,6 +18,11 @@ import type {
     GuidedAnswerTreeSearchResult,
     GuidedAnswersQueryOptions,
     BetaFeatures,
+    GuideFeedback,
+    FeedbackOutcomePayload,
+    SendFeedbackOutcome,
+    FeedbackCommentPayload,
+    SendFeedbackComment,
     SetProductFilters,
     SetComponentFilters,
     ResetFilters
@@ -36,6 +41,9 @@ import {
     UPDATE_LOADING,
     WEBVIEW_READY,
     BETA_FEATURES,
+    GUIDE_FEEDBACK,
+    SEND_FEEDBACK_OUTCOME,
+    SEND_FEEDBACK_COMMENT,
     SET_PRODUCT_FILTERS,
     SET_COMPONENT_FILTERS,
     RESET_FILTERS
@@ -110,3 +118,18 @@ export const resetFilters = (): ResetFilters => ({
 });
 
 export const webviewReady = (): WebviewReady => ({ type: WEBVIEW_READY });
+
+export const guideFeedback = (payload: boolean | null): GuideFeedback => ({
+    type: GUIDE_FEEDBACK,
+    payload
+});
+
+export const sendFeedbackOutcome = (payload: FeedbackOutcomePayload): SendFeedbackOutcome => ({
+    type: SEND_FEEDBACK_OUTCOME,
+    payload
+});
+
+export const sendFeedbackComment = (payload: FeedbackCommentPayload): SendFeedbackComment => ({
+    type: SEND_FEEDBACK_COMMENT,
+    payload
+});