+++ packages/webapp/src/webview/types.ts
@@ -11,6 +11,8 @@ export interface AppState {
     activeGuidedAnswerNode: GuidedAnswerNode[];
     activeGuidedAnswer?: GuidedAnswerTree;
     betaFeatures: boolean;
+    searchResultCount: number;
+    guideFeedback: null | boolean;
     selectedProductFilters: string[];
     selectedComponentFilters: string[];
 }