+++ packages/webapp/src/webview/ui/components/GuidedAnswerNavPath/GuidedAnswerNavPath.tsx
@@ -15,8 +15,11 @@ import { focusOnElement } from '../utils';
 export function GuidedAnswerNavPath(): ReactElement {
     let firstTimeFocus: boolean;
     const nodes = useSelector<AppState, GuidedAnswerNode[]>((state) => state.activeGuidedAnswerNode);
-    const activeNode = nodes[nodes.length - 1];
-    const lastIndex = nodes.length - 1;
+    const isSolved = useSelector<AppState, boolean | null>((state) => state.guideFeedback);
+    let activeNode = nodes[nodes.length - 1];
+    let lastIndex = nodes.length - 1;
+    let lastBlockBorderStyle =
+        isSolved === false ? 'timeline-content-bottom-border-not-solved' : 'timeline-content-bottom-border';
 
     if (activeNode) {
         firstTimeFocus = true;
@@ -36,9 +39,7 @@ export function GuidedAnswerNavPath(): ReactElement {
                         return (
                             <div key={`timeline-block-${i}`} className="timeline-block" role="treeitem">
                                 <button
-                                    className={`timeline-content ${
-                                        i === lastIndex ? 'timeline-content-bottom-border' : ''
-                                    }`}
+                                    className={`timeline-content ${i === lastIndex ? lastBlockBorderStyle : ''}`}
                                     onClick={(): void => {
                                         actions.updateActiveNode(node);
                                     }}>