+++ packages/webapp/src/webview/ui/components/GuidedAnswerNode/Middle/Middle.tsx
@@ -4,6 +4,10 @@ import type { GuidedAnswerNode as GuidedAnswerNodeType } from '@sap/guided-answe
 import '../GuidedAnswerNode.scss';
 import { FocusZone, FocusZoneDirection } from '@fluentui/react-focus';
 import { focusOnElement } from '../../utils';
+import { useSelector } from 'react-redux';
+import { AppState } from '../../../../types';
+import { FeedbackSection } from '../../FeedbackSection/FeedbackSection';
+import NotSolvedMessage from '../../FeedbackSection/NotSolvedMessage/NotSolvedMessage';
 
 let firstTimeFocus = true;
 
@@ -19,8 +23,11 @@ export function Middle(props: {
     activeNode: GuidedAnswerNodeType;
     enhancedBody: ReactElement | undefined | null;
 }): ReactElement {
+    const appState = useSelector<AppState, AppState>((state) => state);
     firstTimeFocus = true;
-    return (
+    return appState.guideFeedback === false ? (
+        <NotSolvedMessage />
+    ) : (
         <div id="middle" className="column">
             <div className="body_container">
                 <header>{props.activeNode.TITLE}</header>
@@ -47,18 +54,22 @@ export function Middle(props: {
                 direction={FocusZoneDirection.bidirectional}
                 isCircularNavigation={true}>
                 <div className="guided-answer__node">
-                    {props.activeNode.EDGES.map((edge, index) => (
-                        <button
-                            role="option"
-                            key={`edge_button${index}`}
-                            className="guided-answer__node__edge"
-                            onClick={(): void => {
-                                actions.selectNode(edge.TARGET_NODE);
-                                focusOnElement('.home-icon');
-                            }}>
-                            {edge.LABEL}
-                        </button>
-                    ))}
+                    {props.activeNode.EDGES.length > 0 ? (
+                        props.activeNode.EDGES.map((edge, index) => (
+                            <button
+                                role="option"
+                                key={`edge_button${index}`}
+                                className="guided-answer__node__edge"
+                                onClick={(): void => {
+                                    actions.selectNode(edge.TARGET_NODE);
+                                    focusOnElement('.home-icon');
+                                }}>
+                                {edge.LABEL}
+                            </button>
+                        ))
+                    ) : (
+                        <FeedbackSection />
+                    )}
                 </div>
             </FocusZone>
         </div>