+++ packages/ide-extension/CHANGELOG.md
@@ -1,5 +1,11 @@
 # sap-guided-answer-extension
 
+## 0.13.0
+
+### Minor Changes
+
+-   2b297ad: solved / not solved feedback interface added
+
 ## 0.12.0
 
 ### Minor Changes