+++ packages/webapp/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap/guided-answers-extension-webapp
 
+## 0.14.0
+
+### Minor Changes
+
+-   2b297ad: solved / not solved feedback interface added
+
+## 0.13.0
+
+### Minor Changes
+
+-   3401231: fix(ga): fix a11y issues in dialog
+
 ## 0.12.0
 
 ### Minor Changes